pyproject.toml
[tool.poetry]
name = "futon"
version = "0.1.0"
description = "Create automated bots that trade for you while you sleep"
authors = ["Rohit Agrawal <rohitagrawalofficialmail@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = ""
repository = "https://github.com/agrawal-rohit/futon"
documentation = "https://github.com/agrawal-rohit/futon"
keywords = [
    "Stocks",
    "Cryptocurrencies",
    "Algorithmic Trading",
    "Backtesting",
    "Trading bots"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
include = ["CHANGELOG.md"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry.dependencies]
python = "^3.7.1"
python_binance = "1.0.10"
bokeh = "2.3.2"
TA_Lib = "0.4.20"
pandas = "1.2.4"
numpy = "*"
binance = "0.3"
Requests = "2.25.1"
tqdm = "4.61.1"
websocket_client = "1.1.0"

[tool.poetry.dev-dependencies]
coverage = "5.3"
flake8 = "*"
pipenv-to-requirements = "*"
pre-commit = "*"
black = "*"
# [tool.poetry.scripts]
# test = "python -m unittest discover"
# coverage = "coverage run --source=futon -m unittest discover && coverage report -m && coverage xml"
